third_party/WebKit/LayoutTests/http/tests/devtools/components/parsed-url-expected.txt

Tests inspector ParsedURL class

Parsing url: http://example.com/?queryParam1=value1&queryParam2=value2#fragmentWith/Many//Slashes
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: /
  queryParams: queryParam1=value1&queryParam2=value2
  fragment: fragmentWith/Many//Slashes
  folderPathComponents: 
  lastPathComponent: 
Parsing url: http://example.com/foo.html?queryParam1=value1@&queryParam2=value2#fragmentWith/Many//Slashes
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: /foo.html
  queryParams: queryParam1=value1@&queryParam2=value2
  fragment: fragmentWith/Many//Slashes
  folderPathComponents: 
  lastPathComponent: foo.html
Parsing url: http://user42:Alina-!$&@example.com/foo/bar.html?queryParam1=value1&queryParam2=value2#fragmentWith/Many//Slashes
  isValid: true
  scheme: http
  user: user42:Alina-!$&
  host: example.com
  port: undefined
  path: /foo/bar.html
  queryParams: queryParam1=value1&queryParam2=value2
  fragment: fragmentWith/Many//Slashes
  folderPathComponents: /foo
  lastPathComponent: bar.html
Parsing url: http://foo@example.com/foo/bar/baz.html?queryParam1=value1&queryParam2=value2#fragmentWith/Many//Slashes
  isValid: true
  scheme: http
  user: foo
  host: example.com
  port: undefined
  path: /foo/bar/baz.html
  queryParams: queryParam1=value1&queryParam2=value2
  fragment: fragmentWith/Many//Slashes
  folderPathComponents: /foo/bar
  lastPathComponent: baz.html
Parsing url: http://example.com//?queryParam1=value1
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: //
  queryParams: queryParam1=value1
  fragment: undefined
  folderPathComponents: /
  lastPathComponent: 
Parsing url: http://example.com//foo.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: //foo.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /
  lastPathComponent: foo.html
Parsing url: http://example.com//foo/bar.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: //foo/bar.html
  queryParams: 
  fragment: undefined
  folderPathComponents: //foo
  lastPathComponent: bar.html
Parsing url: http://example.com/foo//bar.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: /foo//bar.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo/
  lastPathComponent: bar.html
Parsing url: http://example.com//foo//bar.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: //foo//bar.html
  queryParams: 
  fragment: undefined
  folderPathComponents: //foo/
  lastPathComponent: bar.html
Parsing url: http://example.com//foo/bar/baz.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: //foo/bar/baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: //foo/bar
  lastPathComponent: baz.html
Parsing url: http://example.com/foo//bar/baz.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: /foo//bar/baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo//bar
  lastPathComponent: baz.html
Parsing url: http://example.com/foo/bar//baz.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: /foo/bar//baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo/bar/
  lastPathComponent: baz.html
Parsing url: http://example.com///foo/bar/baz.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: ///foo/bar/baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: ///foo/bar
  lastPathComponent: baz.html
Parsing url: http://example.com/foo////bar/baz.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: /foo////bar/baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo////bar
  lastPathComponent: baz.html
Parsing url: http://example.com/foo/bar/////baz.html
  isValid: true
  scheme: http
  user: undefined
  host: example.com
  port: undefined
  path: /foo/bar/////baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo/bar////
  lastPathComponent: baz.html
Splitting url: http://www.chromium.org
  URL: http://www.chromium.org
  Line: undefined
  Column: undefined
Splitting url: http://www.chromium.org:8000
  URL: http://www.chromium.org:8000
  Line: undefined
  Column: undefined
Splitting url: http://www.chromium.org:8000/
  URL: http://www.chromium.org:8000/
  Line: undefined
  Column: undefined
Splitting url: http://www.chromium.org:8000/foo.js:10
  URL: http://www.chromium.org:8000/foo.js
  Line: 9
  Column: undefined
Splitting url: http://www.chromium.org:8000/foo.js:10:20
  URL: http://www.chromium.org:8000/foo.js
  Line: 9
  Column: 19
Splitting url: http://www.chromium.org/foo.js:10
  URL: http://www.chromium.org/foo.js
  Line: 9
  Column: undefined
Splitting url: http://www.chromium.org/foo.js:10:20
  URL: http://www.chromium.org/foo.js
  Line: 9
  Column: 19
